requirements_pi.txt
# requirements_pi.txt
tflite-runtime>=2.13.0
pyaudio>=0.2.13
librosa>=0.10.0
numpy>=1.23.0
RPi.GPIO>=0.7.1
